Move code from core crate into main crate
