+++ Cargo.toml
@@ -6,14 +6,11 @@ authors = [
     "Julian Kniephoff <me@juliankniephoff.com>",
 ]
 
-[workspace]
-
 [profile.dev]
 opt-level = 1
 codegen-units = 4
 
 [dependencies]
-core = { path = "core/" }
 arrayvec = "0.3.20"
 cgmath = "0.12.0"
 env_logger = "0.4.0"
@@ -21,5 +18,6 @@ error-chain = "0.9.0"
 glium = "0.16.0"
 log = "0.3.6"
 num_cpus = "1.2.1"
+rayon = "0.4"
 term-painter = "0.2.3"
 threadpool = "1.3.2"
