--- core/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "core"
-version = "0.1.0"
-authors = [
-    "Lukas Kalbertodt <lukas.kalbertodt@gmail.com>",
-    "Julian Kniephoff <me@juliankniephoff.com>",
-]
-
-[dependencies]
-cgmath = "0.12.0"
-rayon = "0.4"
