--- core/src/lib.rs
@@ -1,6 +1,0 @@
-extern crate cgmath;
-
-pub mod shape;
-pub mod math;
-
-pub use shape::Shape;
