+++ src/app.rs
@@ -5,8 +5,8 @@ use camera::Projection;
 use control::Fly as FlyControl;
 use control::Orbit as OrbitControl;
 use control::{CamControl, KeySwitcher};
-use core::math::*;
-use core::shape::{Shape, Mandelbulb, Sphere};
+use math::*;
+use shape::{Shape, Mandelbulb, Sphere};
 use env::Environment;
 use errors::*;
 use event::{EventResponse, poll_events_with, QuitHandler};
