+++ src/control/mod.rs
@@ -2,8 +2,8 @@ use glium::glutin::Event;
 use glium::glutin::VirtualKeyCode;
 
 use camera::{Camera, Projection};
-use core::math::*;
-use core::Shape;
+use math::*;
+use shape::Shape;
 use event::{EventHandler, EventResponse};
 
 mod fly;
