+++ src/math.rs
