+++ src/shape/util.rs
