+++ src/util/gl.rs
@@ -4,7 +4,7 @@ use std::fs::File;
 use std::io::Read;
 use std::path::Path;
 
-use core::Shape;
+use shape::Shape;
 use errors::*;
